backend/requirements.txt
django>=3.2,<4.0
djangorestframework
psycopg2-binary